Verilog code for single-node time-delay RC
Included are the verilog files used to collect the data for
http://arxiv.org/abs/1411.1398 . This publication describes an
experiment using asynchronous digital logic on a field-programmable gate
array (FPGA) to construct a physical reservoir computer to solve simple
classification problems.

The Verilog code was compiled using Altera's Quartus II software. The
FPGA used in the experiment was an Altera Cyclone IV EP4CE115F29C7N on a
Terasic De2-115 development board. Data was transferred to the computer
using a FTDI FT232H USB chip that was attached to the FPGA's GPIO pins
(see the USB_comm repo for more info).